--- a/idx/memory/find_cache.go
+++ b/idx/memory/find_cache.go
@@ -275,7 +275,7 @@ func (c *FindCache) processInvalidateQueue() {
 			}
 
 			for _, k := range cache.Keys() {
-				matches, err := find((*Tree)(tree), k.(string))
+				matches, err := findOneShot((*Tree)(tree), k.(string))
 				if err != nil {
 					log.Errorf("memory-idx: checking if cache key %q matches any of the %d invalid patterns resulted in error: %s", k, len(reqs), err)
 				}