this implements the proposal from #1930 by trying to make only minimal changes to the core existing find logic, and use channels to "yield" (as in python) values out of that function.
For an alternative approach, see #1936 

review this commit by commit.