@Dieterbe 2020-10-30
I wouldn't want the new code to slow down finds, seems they have gotten perhaps a tad slower. needs a bit more testing.
perhaps there's some small tweaks to speed it up a bit. i can look into that later.
```
benchstat master.txt after.txt
name                                    old time/op    new time/op    delta
Find/partitioned-2                        95.4µs ± 0%   114.1µs ± 0%   ~     (p=1.000 n=1+1)
Find/unPartitioned-2                      32.3µs ± 0%    32.2µs ± 0%   ~     (p=1.000 n=1+1)
Concurrent4Find/partitioned-2             84.0µs ± 0%    96.9µs ± 0%   ~     (p=1.000 n=1+1)
Concurrent4Find/unPartitioned-2           61.2µs ± 0%    82.9µs ± 0%   ~     (p=1.000 n=1+1)
Concurrent8Find/partitioned-2             80.2µs ± 0%    96.5µs ± 0%   ~     (p=1.000 n=1+1)
Concurrent8Find/unPartitioned-2           64.4µs ± 0%    51.9µs ± 0%   ~     (p=1.000 n=1+1)
ConcurrentInsertFind/partitioned-2        81.4µs ± 0%    99.2µs ± 0%   ~     (p=1.000 n=1+1)
ConcurrentInsertFind/unPartitioned-2      53.1µs ± 0%    53.0µs ± 0%   ~     (p=1.000 n=1+1)
TagFindSimpleIntersect/partitioned-2      1.98ms ± 0%    1.54ms ± 0%   ~     (p=1.000 n=1+1)
TagFindSimpleIntersect/unPartitioned-2    1.61ms ± 0%    1.65ms ± 0%   ~     (p=1.000 n=1+1)
TagFindRegexIntersect/partitioned-2       65.2ms ± 0%    66.2ms ± 0%   ~     (p=1.000 n=1+1)
TagFindRegexIntersect/unPartitioned-2     64.2ms ± 0%    62.5ms ± 0%   ~     (p=1.000 n=1+1)

name                                    old alloc/op   new alloc/op   delta
Find/partitioned-2                        30.2kB ± 0%    31.5kB ± 0%   ~     (p=1.000 n=1+1)
Find/unPartitioned-2                      19.7kB ± 0%    19.9kB ± 0%   ~     (p=1.000 n=1+1)
Concurrent4Find/partitioned-2             30.1kB ± 0%    31.5kB ± 0%   ~     (p=1.000 n=1+1)
Concurrent4Find/unPartitioned-2           19.7kB ± 0%    19.9kB ± 0%   ~     (p=1.000 n=1+1)
Concurrent8Find/partitioned-2             30.2kB ± 0%    31.5kB ± 0%   ~     (p=1.000 n=1+1)
Concurrent8Find/unPartitioned-2           19.7kB ± 0%    19.9kB ± 0%   ~     (p=1.000 n=1+1)
ConcurrentInsertFind/partitioned-2        30.2kB ± 0%    31.6kB ± 0%   ~     (p=1.000 n=1+1)
ConcurrentInsertFind/unPartitioned-2      19.7kB ± 0%    19.9kB ± 0%   ~     (p=1.000 n=1+1)
TagFindSimpleIntersect/partitioned-2      29.6kB ± 0%    29.6kB ± 0%   ~     (p=1.000 n=1+1)
TagFindSimpleIntersect/unPartitioned-2    4.62kB ± 0%    4.62kB ± 0%   ~     (p=1.000 n=1+1)
TagFindRegexIntersect/partitioned-2       1.63MB ± 0%    1.63MB ± 0%   ~     (p=1.000 n=1+1)
TagFindRegexIntersect/unPartitioned-2      513kB ± 0%     493kB ± 0%   ~     (p=1.000 n=1+1)

name                                    old allocs/op  new allocs/op  delta
Find/partitioned-2                           106 ± 0%       120 ± 0%   ~     (p=1.000 n=1+1)
Find/unPartitioned-2                        46.0 ± 0%      48.0 ± 0%   ~     (p=1.000 n=1+1)
Concurrent4Find/partitioned-2                105 ± 0%       119 ± 0%   ~     (p=1.000 n=1+1)
Concurrent4Find/unPartitioned-2             46.0 ± 0%      48.0 ± 0%   ~     (p=1.000 n=1+1)
Concurrent8Find/partitioned-2                105 ± 0%       119 ± 0%   ~     (p=1.000 n=1+1)
Concurrent8Find/unPartitioned-2             46.0 ± 0%      48.0 ± 0%   ~     (p=1.000 n=1+1)
ConcurrentInsertFind/partitioned-2           106 ± 0%       120 ± 0%   ~     (p=1.000 n=1+1)
ConcurrentInsertFind/unPartitioned-2        46.0 ± 0%      48.0 ± 0%   ~     (p=1.000 n=1+1)
TagFindSimpleIntersect/partitioned-2         197 ± 0%       197 ± 0%   ~     (all equal)
TagFindSimpleIntersect/unPartitioned-2      36.0 ± 0%      36.0 ± 0%   ~     (all equal)
TagFindRegexIntersect/partitioned-2        20.5k ± 0%     20.5k ± 0%   ~     (p=1.000 n=1+1)
TagFindRegexIntersect/unPartitioned-2      4.23k ± 0%     4.06k ± 0%   ~     (p=1.000 n=1+1)
```