@Dieterbe 2020-11-06
didn't get consistent results out of the benchmarks, but i think this should be good enough to merge, we can deploy to our internal instances and see if it needs further improvements